+++ crates/miden-objects/src/block/partial_account_tree.rs
@@ -1,12 +1,10 @@
 use miden_crypto::merkle::SmtLeaf;
 
-use crate::{
-    Word,
-    account::AccountId,
-    block::{AccountTree, AccountWitness},
-    crypto::merkle::PartialSmt,
-    errors::AccountTreeError,
-};
+use crate::Word;
+use crate::account::AccountId;
+use crate::block::{AccountTree, AccountWitness};
+use crate::crypto::merkle::PartialSmt;
+use crate::errors::AccountTreeError;
 
 /// The partial sparse merkle tree containing the state commitments of accounts in the chain.
 ///
@@ -161,12 +159,12 @@ impl PartialAccountTree {
         // Note that if the leaf is empty, that's fine. It means it is tracked by the partial SMT,
         // but no account ID is inserted yet.
         // Also note that the multiple variant cannot occur by construction of the tree.
-        if let Ok(SmtLeaf::Single((existing_key, _))) = self.smt.get_leaf(&key) {
-            if key != existing_key {
-                return Err(AccountTreeError::DuplicateIdPrefix {
-                    duplicate_prefix: account_id.prefix(),
-                });
-            }
+        if let Ok(SmtLeaf::Single((existing_key, _))) = self.smt.get_leaf(&key)
+            && key != existing_key
+        {
+            return Err(AccountTreeError::DuplicateIdPrefix {
+                duplicate_prefix: account_id.prefix(),
+            });
         }
 
         self.smt